refactored code to be a bit more idomatic rust
